13175 Move POS asset tests to new file
